edge-service/requirements.txt
Flask
pyserial
mariadb
